+++ src/modalVariant/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import SpidIcoCircleLbUrl from './/../shared/svgs/spid-ico-circle-lb.svg';
-import SpidIcoCircleBbUrl from './/../shared/svgs/spid-ico-circle-lb.svg';
+import SpidIcoCircleLbUrl from './spid-ico-circle-lb.svg';
+import SpidIcoCircleBbUrl from './spid-ico-circle-lb.svg';
 
 import { getTranslationFn } from '../shared/i18n';
 import {
